Develop

+++ src/main/java/com/talexframe/frame/core/modules/network/connection/app/addon/cache/redis/ReceiverCacheRedisAddon.java
@@ -33,7 +33,7 @@ public class ReceiverCacheRedisAddon extends ReceiverAddon {
 
         super("ReceiverMethod", new ReceiverAddonType[] { ReceiverAddonType.METHOD_APP });
 
-        super.priority = ReceiverAddonPriority.MOST_HIGHEST;
+        super.priority = ReceiverAddonPriority.LOW;
 
     }
 
@@ -44,15 +44,29 @@ public class ReceiverCacheRedisAddon extends ReceiverAddon {
 
         if( tRedisCache == null ) return true;
 
+        Class<?> returnType = methodAppReceiver.getMethod().getReturnType();
+
+        if( returnType.isAssignableFrom(Void.TYPE) ) {
+
+            log.warn("[RedisCache] 方法返回类型为Void, 不支持缓存! ");
+
+            return true;
+
+        }
+
         RedisTemplate<String, Object> template = redis.getConfig().getRedisTemplate();
 
         ValueOperations<String, Object> vo = template.opsForValue();
 
         if ( !tRedisCache.delete() ) {
 
-            String key = getRedisCacheKey( tRedisCache, wr.getRequest().getRequestURI(), methodAppReceiver.getParams() );
+            String key = getRedisCacheKey( tRedisCache, wr.getRequest().getRequestURI(), wr.getParams() );
 
-            wr.returnData(vo.get(key));
+            Object obj = vo.get(key);
+
+            if( obj == null ) return true;
+
+            wr.returnData(obj);
 
             log.info("[RedisCache] Cache Hit!");
 
@@ -65,7 +79,7 @@ public class ReceiverCacheRedisAddon extends ReceiverAddon {
     }
 
     @Override
-    public void onPostAddParam(MethodAppReceiver methodAppReceiver, WrappedResponse wr, Object methodReturn) {
+    public void onPostInvokeMethod(MethodAppReceiver methodAppReceiver, WrappedResponse wr, Object methodReturn) {
 
         TRedisCache tRedisCache = methodAppReceiver.getMethod().getAnnotation(TRedisCache.class);
 
@@ -81,7 +95,17 @@ public class ReceiverCacheRedisAddon extends ReceiverAddon {
 
         }
 
-        String key = getRedisCacheKey( tRedisCache, wr.getRequest().getRequestURI(), methodAppReceiver.getParams() );
+        Class<?> returnType = methodAppReceiver.getMethod().getReturnType();
+
+        if( returnType.isAssignableFrom(Void.TYPE) ) {
+
+            log.warn("[RedisCache] 方法返回类型为Void, 不支持缓存! ");
+
+            return;
+
+        }
+
+        String key = getRedisCacheKey( tRedisCache, wr.getRequest().getRequestURI(), wr.getParams() );
 
         if ( tRedisCache.delete() ) {
 
@@ -126,7 +150,10 @@ public class ReceiverCacheRedisAddon extends ReceiverAddon {
 
                 int index = Integer.parseInt(key.substring(8, 9));
 
-                if ( index + 1 > params.size() ) {
+                if ( index > params.size() ) {
+
+                    log.warn("[RedisCache] @RedisCache #params index out of range");
+                    log.warn("[RedisCache] For param index: {} with params: {}", index, params);
 
                     throw new RuntimeException("params index out of range");
